Update container values

+++ src/main/java/com/labsynch/labseer/service/AssayService.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Service;
 @Service
 public interface AssayService {
     
-	int renameBatchCode(String oldCode, String newCode, String modifiedByUser);
+	Long renameBatchCode(String oldCode, String newCode, String modifiedByUser);
 
 }
